Add files via upload
Added method to check for duplicate entries to the tree after deletion for testing. To make sure the original copy of the node that has been moved up the tree is deleted. Unfortunately I could only add the appropriate methods to the BinaryTreeLUT class, because they need to use the BSTreeNode object, and importing into the test file wouldn't work. So if you want to use it you have to add the appropriate methods (which are in BinaryTreeLUT_Printable) to your BinaryTreeLUT class. You should probably then delete them again before you hand in. I'm not sure if Jens would notice multiple people having the same strange method added to their submission.