Fun guaranteed
These were kind of hard to test thoroughly because of the weird printing, but hopefully this is of some help. I've generally tested for success by making sure the key that was supposed to be removed is no longer in the string representation, and I'm not 100% sure that that doesn't leave a loophole to pass these tests and still not be working properly.