No Fun Involved, Sorry
Please note: there are some valid solutions that will fail these tests due to the order of insertions and collisions. If A and B both hash to the same value, then having [... A, B...] in your table is just as legitimate as having [... B, A, ...] but I'm only checking for one of them.
Also note that I can't properly implement fun stuff in JUnit, simply because of the way it works. Sorry about that.